editoast: move `schemas!` to `editoast_common`
