+++ editoast/editoast_common/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod rangemap_utils;
+pub mod schemas;
