+++ editoast/src/map/bounding_box.rs
@@ -9,7 +9,7 @@ use utoipa::ToSchema;
 
 use crate::error::Result;
 
-crate::schemas! {
+editoast_common::schemas! {
     Zone,
     BoundingBox,
 }
