+++ editoast/src/views/documents.rs
@@ -25,7 +25,7 @@ crate::routes! {
     }
 }
 
-crate::schemas! {
+editoast_common::schemas! {
     NewDocumentResponse,
 }
