+++ editoast/src/views/electrical_profiles.rs
@@ -40,7 +40,7 @@ crate::routes! {
     }
 }
 
-crate::schemas! {
+editoast_common::schemas! {
     LightElectricalProfileSet,
     ElectricalProfile,
     ElectricalProfileSetData,
