+++ editoast/src/views/v2/path.rs
@@ -25,7 +25,7 @@ crate::routes! {
     pathfinding::routes(),
 }
 
-crate::schemas! {
+editoast_common::schemas! {
     pathfinding::schemas(),
     properties::schemas(),
 }
