+++ editoast/src/views/v2/path/pathfinding.rs
@@ -42,7 +42,7 @@ crate::routes! {
     },
 }
 
-crate::schemas! {
+editoast_common::schemas! {
     PathfindingInput,
     PathfindingResult,
     TrackRange,
