+++ editoast/src/views/v2/train_schedule.rs
@@ -54,7 +54,7 @@ crate::routes! {
     },
 }
 
-crate::schemas! {
+editoast_common::schemas! {
     Distribution,
     TrainScheduleBase,
     TrainScheduleForm,
